Type React Helmet dummy flow
