+++ react_on_rails/spec/dummy/client/app/startup/ReactHelmetApp.client.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { HelmetProvider } from '@dr.pogodin/react-helmet';
+import ReactHelmet, { type ReactHelmetProps } from '../components/ReactHelmet';
+
+const ReactHelmetApp = (props: ReactHelmetProps) => (
+  <HelmetProvider>
+    <ReactHelmet {...props} />
+  </HelmetProvider>
+);
+
+export default ReactHelmetApp;
