+++ react_on_rails/spec/dummy/client/app/startup/ReactHelmetApp.server.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { HelmetProvider, type HelmetDataContext } from '@dr.pogodin/react-helmet';
+import { renderToString } from 'react-dom/server';
+import type { RailsContext, ServerRenderResult } from 'react-on-rails/types';
+import ReactHelmet, { type ReactHelmetProps } from '../components/ReactHelmet';
+
+const ReactHelmetApp = (props: ReactHelmetProps, _railsContext: RailsContext): ServerRenderResult => {
+  const helmetContext: HelmetDataContext = {};
+  const componentHtml = renderToString(
+    <HelmetProvider context={helmetContext}>
+      <ReactHelmet {...props} />
+    </HelmetProvider>,
+  );
+  const { helmet } = helmetContext;
+
+  return {
+    renderedHtml: {
+      componentHtml,
+      title: helmet?.title?.toString() || '',
+    },
+  };
+};
+
+export default ReactHelmetApp;
