+++ react_on_rails/spec/dummy/client/app/startup/ReactHelmetAppBroken.client.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { HelmetProvider } from '@dr.pogodin/react-helmet';
+import ReactHelmet, { type ReactHelmetProps } from '../components/ReactHelmet';
+
+const ReactHelmetAppBroken = (props: ReactHelmetProps) => (
+  <HelmetProvider>
+    <ReactHelmet {...props} />
+  </HelmetProvider>
+);
+
+export default ReactHelmetAppBroken;
